Removing lengthy redis checks
Looking for drains is too expensive as a regular operation. In
production it may take up to two minutes just to run all the checks and
Redis is what makes it slow. Tests have shown that no drain currently on
a production node appears to be orphaned in either ETS or Redis, but
always both. As such, we drop this lookup.