Allowing write concurrency on logplex_stats
Logplex_stats is a table that receives a very large amount of writes for
one big read every minute. We should be able to generally benefit by
reducing contention on resources given pretty much all drains write to
that table, at the cost of a slightly longer blocking time when reading.