Move web responses before IO logging calls
With IO being blocking for individual processes due to Erlang's IO
protocol and logplex using io:format/2 to log information, it is
possible that a node that does a lot of logging has bad tail latencies
on its API as reported by issues #49 and #51 on github.

This quickfix, pending a rewrite of the logging system to be
non-blocking and load-shedding, moves the logging outside of the
critical path for part of the requests as a whole. Some requests, such
as token creation for channels (POST /v2/channels/(\\d+)/tokens) still
contain logs in said critical path and will only see minor improvements.