Force hibernation on drain reconnect
When there's a timer being set for a reconnection, we force hibernation
in order to do a fullsweep GC of the drain processes.

This might incur a certain cost for very busy-but-disconnected
processes, forcing a short pause, but the backoff timers for
reconnections will act as rate limiters on this.