Making counters explicit for buffer size
The logplex_msg_buffer module is used extensively by drain processes
that buffer request and need to be the least blocking possible under
heavy load. The current implementation would recalculate the entire
queue length on every call, which became both time consuming and CPU
intensive when the buffer was full, which happens when you have to count
lengths even more often.

This patch makes it so that we have an explicit counter for the buffer
so that we don't need to recalculate it all the time, lowering the
contention for runtime for a given process.

The module includes conversion clauses for all functions part of the API
so that the code can be hot-reloaded without stopping, and just adapt to
the new format.