Adding hibernation timeouts on drains
An inactive drain or buffer (Receives no request from the outside world)
should be sent to hibernation in order to trigger a full-sweep GC,
compact the memory of the process, and reduce the overall load of the
system, and possibly reducing memory fragmentation of the VM at the cost
of slightly more CPU when it triggers.

The timeout is implemented using the gen_fsm timeout option, which
automatically resets timeout timers when a message is processed by the
process. This should allow to generally catch any kind of inactivity and
force hibernation of the processes.

Note: it is not yet known if the timeout value of 5 seconds or the
amount of timers setup/cancellations will have an impact of any
significance on an active system or not. The values may need to be
tweaked or the effort redirected towards manual GC if refc binaries keep
on hogging the memory after this.