Moving logging to lager
The current logplex version shows a point of contention for logs through
using io:format/2. Although it is unlikely lager will help a lot with
it given we don't log directly to disk (and this is where it shines in
comparison to other logging engines), it's worth trying to see if things
are improving with it.

Custom log formats are used to make sure the production log format
remains 100% identical to the former one. They will, however, be
different during test runs because no specific care has been taken to
make the lager config be compatible in test cases.