Raising send timeout to 5s on HTTP(s) drains
When network or drains show bad behaviour temporarily, low timeouts we
currently have (1 second) ends up killing connections and raising the
retry count of frames. When massive losses are seen, it makes it
difficult to put the blame on logplex's speed at sending logs, or the
drains consumption (or network).

By raising the timeout a bit, we should reduce the reconnection rate and
at the same time make it harder to blame logplex (as an individual node)
for the problems.

This should not have a super significant impact on the drop rate,
however, but possibly a noticeable one.