Logplex Shard commits do not update URL cache
The cache that can be obtained with logplex_shard:urls() is not updated
when committing a replacement shard successfully. This can lead to
confusing subsequent calls if nodes have been removed in ETS tables but
the operator still uses urls/0 to create the new shard list.

This patch makes sure that the url cache is update using the
?CURRENT_WRITE_MAP table, which should be similar to the read pool given
the current implementation.