Close even active drains after they've been around too long.
This will avoid bad behaviour around load balancers where more
reliable nodes tend to accumulate too many connections over time as
the other ones drop out.

We do this by piggybacking on the existing timer we're using for idle
timeouts, so we should probably rename that timer.

The original design also called for randomizing the total timeout, but
this implementation just piggybacks on the fuzzing already present in
the idle timer. This may not be sufficient fuzzing; there's a check
every fuzzed five minutes to see whether it's gone over the 5 hour
limit, but it would probably better if the 5 hour limit itself were
fuzzed.