Simply restart close timer if it triggers in sending.
This should be more graceful, but it means that sometimes we'll have
active drains that last longer than the max TTL by a several
increments of the idle timer. I don't think this is a big problem
since the max TTL is a "soft maximum" intended to improve the load
balancing properties rather than something that needs strict adherence.