resolve replace_shards issue when shard is down
The issue was caused by the `timer:sleep` call in the `init` callback.
Under normal conditions the `init` callback is called synchronously and
`start_link` will not return until `init` returns. The `timer:sleep`
call was causing the supervisor to lock for 1s waiting for the init to
finish. This meant each time the downed shard writer (10 of them)
failed, it would post-pone even the new shard writers from being
setup.

In this case `proc_lib` is used, so it is possible to cheat and ack the
init before even attempting to connect to redis. The supervisor is then
able to process other child starts, while the writer process is still
figuring out if it can even reach the redis instance.