remove report_handler and add an alarm_handler
The `logplex_report_handler` was not being used. It was most likely too
noisy in terms of paging. It is now replaced by a
`logplex_alarm_handler` module that pages when alarms are raised within
the system.

Alarms can be raised via the exported `set_alarm/1` and
cleared via `clear_alarm/1`. The interface mimics that of OTP
`alarm_handler`, and in fact just delegates to the stock `alarm_handler`
when appropriate.