support triggering and resolving alarms in pagerduty
Alarms can be triggered multiple times, each time the alarm is triggered
it will use the pagerduty api to trigger a page. This will not cause
multiple pages for the same incident, instead pagerduty will track each
page in the page log for each unique AlarmId.

Once an alarm is resolve the page will not be cleared until all stored
alarms are also cleared. For example an alarm raised 4 times will need
to be cleared 4 times before the page is finally resolved.