uses folsom and folsom_cowboy for realtime metrics
The `logplex_realtime` has been gutted and replaced by a couple calls to
`folsom_metrics`.

The `ehmon` dependency has been dropped in favour of adding
`folsom_cowboy` and allowing the metrics to be fetched via a locally
listening port.

In order to fascilitate the addition of `folsom_cowboy`, the `cowboy`
dependency has been updated to the latest 1.0.x cowboy branch. It was
necessary to copy over the now deprecated `cowboy_client` to logplex
src. This should really be replaced by `gun` at some point.