fix failing shrink tests
The shrink test incorrectly asserted the old behaviour of resizing the
queue as soon as socket is connected. Once the test was fixed I
discovered and fixed an issue in the logplex_http_drain code that still
set the service level back to normal as soon as the socket is connected
instead of when a successfull HTTP request is sent.