tracks active http drain counts
This commit flips the realtime metrics on its head. Additional changes
will be required to SHH in order to continue parsing the existing
metrics. Folsom allows counters to be incremented and decremented. This
matches the behaviour of a gauge in SHH terms. The existing
monotonically increasing metrics have been switch to a `meter_reader`
folsom type (support must be added in SHH). Folsom counters should be
considered gauges. The `logplex_realtime` API has been updated to be
more clear in this regard.