ensure tokens are deleted everywhere
This commit ensures that tokens are removed on all nodes. Prior to this
commit the token would be deleted from the local node that received a
channel delete API call. Because the token delete command is never
issued through redis to the config master, the token is never cleaned
up on the other nodes. Furthermore, on a system reboot the orphaned
token will be resynced down from the config master to the local node.

This commmit also removes the completely unused destroy method.