use data-volumes for deps & ebin
This docker-compose configuration isolates the deps and ebin directories
into a base image. This base image is run as a data volume by the
logplex service and a new compile service.

```
docker-compose run compile
```

can be used to quickly rebuild the logplex source. To recompile
everything from scratch (including deps) use:

```
docker-compose run compile make all
```

The `Makefile` was added as a workaround to some kind of pseudo-tty bug
with rebar. Running the `rebar compile skip_deps=true` commands directly
Wwould result in rebar hanging. The inclusion of the `Makefile` or some
bash script to execute fixes that issue.

The next task will be to create docker-compose service that connects an
eshell to a running logplex service.