assert max_depth behaviour
Because sessions are resused by default the test issues a stop/start to
the ssl application in order to flush that cache.

The `{verify, verify_peer}` was altered to `{verify_fun, ...}` to allow
easier debugging in the test suite. The supplied fun, is literally
copied from OTP source and is the same fun that would be substituted for
`verify_peer`.