Require API URLs
We need to ensure the API URLs for v2 and v3 API are configured. The v3
API requires them in order to redirect a client with a full URL. The v2 API
would return only a path without a hostname. Note this may change how a client
handles redirects.

The new default sets them to:

* v2: `http://localhost:8001`
* v3: `http://localhost:8002`

The environment variables `LOGPLEX_API_ENDPOINT_URL` and
`LOGPLEX_HTTP_V3_URL` are now required to be set, otherwise the above
mentioned defaults are returned.