Small tweaks (2) to get docker-compose working locally
As a newcomer to the logplex project, I attempted to get the Docker based dev environment up and running. I ran into two issues connecting to the logplex container.

First, when I initially tried to connect to the logplex container per the README, the connection failed, due to a missing file called `keys.sh’. This file ended up being the logplex.env file. I added commands to the dockerfile to create the file that is expected.

Second, I discovered that erlang is very particular wrt how host/nodes are named. The names set for the logplex `domainname` and `instance name` in the `docker-compose.yml` file didn’t match up with the running logplex container. I fixed this by replacing three values in the `docker-compose.yml` file with environment variables, and adding a small shell script that sets these variables before the container is built. See the README changes.